int: Ensure we stringify Booleans as 0 or 1
Python 3.8 removes the tp_str for various built-in types, so we would
print Boolean values as their repr (for example dbus.Boolean(True)),
which is a regression. Print them as 0 or 1 instead, which was the
historical behaviour (arguably False or True would be better, but
that would be a behaviour change).

Signed-off-by: Simon McVittie <smcv@collabora.com>